refactor: merge list.h{,pp}
